fix: Enable running plan/destroy when helmfile_release_set kubeconfig depends on another, missing resource
The provider was running helmfile-diff with the default kubeconfig, not one from helmfile_releaset_set.kubeconfig nor helmfile_release_set.environment_variables.KUBECONFIG when the kubeconfig is
computed/sourced from another terraform resource that is not created yet. This is a common scenario when the provider is used with e.g. terraform-provider-eksctl that creates a K8s cluster and outputs kubeconfig path to be used by other resources.

This change fixes that, by ignoring error when kubeconfig is set to empty, which means that either the user is inteded to use system kubeconfig, or terraform has not yet created the resource required to fulfill the kubeconfig path.

In the future, we may better introduce a dummy kubeconfig as the default, so that the system kubeconfig will not be used unexpectedly.