Always use the absolute path of helmfile_release_set.kubeconfig
This turned out to be more intuitive when you use sub-helmfiles, due to that each sub-helmfile uses the KUBECONFIG relative to it by the nature of Helmfile.

Since this commit, any helmfile command run by this provider can be treated as `KUBECONFIG=$(pwd)/<helmfile_release_set kubecondfig attr> helmfile blah`.